### 2025-07-20 #8619 @warrickball
Hi @vicky1997, thanks for your submission! All the suitable JOSS editors are currently working at capacity so I'm going to "waitlist" this review until an editor with the relevant expertise is available to take it on. Unfortunately at the moment this will probably take at least 4 months. Sorry, and thanks for your patience!

In the meantime, could you comment on the gap in the commit history between 11 Oct 2023 and 26 Jun? It looks like basically the entire codebase was overwritten.

Also, I noticed some citations in the paper are rendering like "(Stammler & Birnstiel (2022))". This is because you've coded the citation as `(@Stammler2022)` rather than `[@Stammler2022]` (note the square brackets), which will render correctly as "(Stammler & Birnstiel 2022)". You've also presumably taken the BibTeX entries from NASA ADS, which uses macros like `\apj` for journal names. Our pipeline doesn't recognise these and, as a multi-disciplinary journal, we require journal names to be written out in full, or at least enough that the domain is clear (e.g. "The Astrophysical Journal").

None of this will affect your place in our submission queue but we might as well handle it now.
