+++ paper/paper.bib
@@ -1,7 +1,7 @@
 @ARTICLE{Stammler2022,
        author = {{Stammler}, S. M. and {Birnstiel}, Tilman},
         title = "{DustPy: A Python Package for Dust Evolution in Protoplanetary Disks}",
-      journal = {\apj},
+      journal = {The Astrophysical Journal},
      keywords = {Protoplanetary disks, Astronomy software, Astronomical simulations, Circumstellar dust, Planet formation, Planetesimals, 1300, 1855, 1857, 236, 1241, 1259, Astrophysics - Earth and Planetary Astrophysics, Astrophysics - Instrumentation and Methods for Astrophysics},
          year = 2022,
         month = aug,
@@ -20,7 +20,7 @@ archivePrefix = {arXiv},
 @ARTICLE{Drazkowska2013,
        author = {{Dr{\k{a}}{\.z}kowska}, J. and {Windmark}, F. and {Dullemond}, C.~P.},
         title = "{Planetesimal formation via sweep-up growth at the inner edge of dead zones}",
-      journal = {\aap},
+      journal = {Astronomy & Astrophysics},
      keywords = {accretion, accretion disks, circumstellar matter, protoplanetary disks, planets and satellites: formation, methods: numerical, Astrophysics - Earth and Planetary Astrophysics},
          year = 2013,
         month = aug,
@@ -38,7 +38,7 @@ archivePrefix = {arXiv},
 @ARTICLE{Robinson2024,
        author = {{Robinson}, A. and {Booth}, Richard A. and {Owen}, James E.},
         title = "{Introducing CUDISC: a 2D code for protoplanetary disc structure and evolution calculations}",
-      journal = {\mnras},
+      journal = {Monthly Notices of the Royal Astronomical Society},
      keywords = {methods: numerical, protoplanetary discs, stars: pre-main-sequence, Astrophysics - Earth and Planetary Astrophysics, Astrophysics - Instrumentation and Methods for Astrophysics},
          year = 2024,
         month = apr,
@@ -56,7 +56,7 @@ archivePrefix = {arXiv},
 @ARTICLE{Zsom2008,
        author = {{Zsom}, A. and {Dullemond}, C.~P.},
         title = "{A representative particle approach to coagulation and fragmentation of dust aggregates and fluid droplets}",
-      journal = {\aap},
+      journal = {Astronomy & Astrophysics},
      keywords = {accretion, accretion disks, methods: statistical, planets and satellites: formation, ISM: dust, extinction, Astrophysics},
          year = 2008,
         month = oct,
@@ -74,7 +74,7 @@ archivePrefix = {arXiv},
 @ARTICLE{2007A&A...461..215O,
        author = {{Ormel}, C.~W. and {Spaans}, M. and {Tielens}, A.~G.~G.~M.},
         title = "{Dust coagulation in protoplanetary disks: porosity matters}",
-      journal = {\aap},
+      journal = {Astronomy & Astrophysics},
      keywords = {ISM: dust, extinction, planetary systems: formation, planetary systems: protoplanetary disks, accretion, accretion disks, Astrophysics},
          year = 2007,
         month = jan,
@@ -109,7 +109,7 @@ archivePrefix = {arXiv},
 @ARTICLE{Price,
[…]
