+++ paper/paper.md
@@ -34,7 +34,7 @@ bibliography: paper.bib
 
 Modelling dust coagulation and evolution is an essential part of understanding planet formation and protoplanetary disks. The evolution of micrometer-sized dust grains to millimeter-sized grains sets the tone for planetesimal formation. Dust evolution is inextricably linked with the formation of substructures in disks. Dust dynamics is also of prime importance when understanding protoplanetary disk chemistry. Therefore very important to understand in detail the processes involved in dust growth and dynamics and how they influence different aspects of disk dynamics and planet growth. 
 
-Dust coagulation is generally modelled with the Smoluchowski equation, an integro-differential equation. Some examples of open source codes that solve the Smoluchowski equation are `dustpy` (@Stammler2022), a 1D code to simulate gas and dust evolution and `cuDisc`(@Robinson2024), a 2D code modelling dust coagulation and disk evolution. Both of these codes are Eulerian in nature, i.e. they follow volume rather than mass. The drawback with these methods is that it is difficult to track histories of dust particles. Furthermore, adding a property to track adds a dimension to solving the Smoluchowski Equation (@Stammler2017). Lagrangian methods can solve this issue partly by allowing us to track histories of particles. `PHANTOM` (@Price), is an open source Smoothed Particle Hydrodynamics (SPH) code that models models dust growth using a monodisperse growth or a 'single-size' approximation and do not solve the Smoluchowski equation for dust growth (@Vericel2021). 
+Dust coagulation is generally modelled with the Smoluchowski equation, an integro-differential equation. Some examples of open source codes that solve the Smoluchowski equation are `dustpy` [@Stammler2022], a 1D code to simulate gas and dust evolution and `cuDisc`[@Robinson2024], a 2D code modelling dust coagulation and disk evolution. Both of these codes are Eulerian in nature, i.e. they follow volume rather than mass. The drawback with these methods is that it is difficult to track histories of dust particles. Furthermore, adding a property to track adds a dimension to solving the Smoluchowski Equation [@Stammler2017]. Lagrangian methods can solve this issue partly by allowing us to track histories of particles. `PHANTOM` [@Price], is an open source Smoothed Particle Hydrodynamics (SPH) code that models models dust growth using a monodisperse growth or a 'single-size' approximation and do not solve the Smoluchowski equation for dust growth [@Vericel2021]. 
 
 Alternatively one can model dust coagulation by performing Monte Carlo simulations. Monte Carlo methods are well suited for stochastic processes and that makes it a suitable method for dust coagulation. The representative particle approach is Lagrangian in nature, meaning we track the particles and hence their histories. We can add properties to the particles that can be tracked without much computational complexity. This advantage is very useful when we want to combine dust evolution and chemistry, where we want to look at the histories of particles and their compositions when chemical models are included. We point the reader to @Vaikundaraman2025 for an example of tracking composition/chemistry using `mcdust` to investigate the carbon depletion puzzle in the inner Solar System.
 
@@ -43,7 +43,7 @@ Including dust growth/dynamics in hydrodynamic simulations of protoplanetary dis
 Currently, there is no available open source code that models dust coagulation using Monte Carlo methods and such a code would be very helpful as an alternative way for modelling dust coagulation.
 
 # Benchmark
-We compare a run of our code with a run from the open source 1D dust coagulation code `dustpy`(@Stammler2022). The parameters used for the simulations are listed in the table below. The gas evolution in both the codes was switched off and the codes were run with a static gas background to exclude any differences in gas treatment that might influence the outcome of dust evolution.
[…]
