+++ paper/paper.md
@@ -60,7 +60,7 @@ We ran a simulation for 10000 years. We show below the dust surface density $\si
 ![](Figures/dustpyvsmcdust.png){width=75%}
 
 
-The image above shows the radial distrbution of It is evident that both the codes have similar overall outcomes but they do have certain differences. The most striking one is that `mcdust` does not provide coverage of the regions of parameter space that do not include sufficiently high fraction of the dust mass. This holds true for all Monte Carlo based codes. But with higher resolution simulation this issue can be overcome. The other important factor is that `mcdust` does not face the issue of artificially sped-up growth that `dustpy` and other Smoluchowski equation based codes tend to encounter. And the 2D r-z structure of `mcdust` also helps us to investigate processes like sedimentation driven coagulation [@Drazkowsa2013] that are not usually seen in 1D simulations like `dustpy`. This can be seen in image at around 50 AU where `mcdust` has larger surface densities higher masses when compared to `dustpy`. For a more detailed discussion of the differences between the two approaches to dust growth, i.e., the Monte Carlo method and the Smoluchowski equation approach we point the reader to @Drazkowska2014.
+The image above shows the radial distrbution of It is evident that both the codes have similar overall outcomes but they do have certain differences. The most striking one is that `mcdust` does not provide coverage of the regions of parameter space that do not include sufficiently high fraction of the dust mass. This holds true for all Monte Carlo based codes. But with higher resolution simulation this issue can be overcome. The other important factor is that `mcdust` does not face the issue of artificially sped-up growth that `dustpy` and other Smoluchowski equation based codes tend to encounter. And the 2D r-z structure of `mcdust` also helps us to investigate processes like sedimentation driven coagulation [@Drazkowska2013] that are not usually seen in 1D simulations like `dustpy`. This can be seen in image at around 50 AU where `mcdust` has larger surface densities higher masses when compared to `dustpy`. For a more detailed discussion of the differences between the two approaches to dust growth, i.e., the Monte Carlo method and the Smoluchowski equation approach we point the reader to @Drazkowska2014.
 
 # Acknowledgments
 The authors acknowledge funding from the European Union under the European Union’s Horizon Europe Research \& Innovation Programme 101040037 (PLANETOIDS). Views and opinions expressed are however those of the author(s) only and do not necessarily reflect those of the European Union or the European Research Council. Neither the European Union nor the granting authority can be held responsible for them.
