+++ docs/source/dustcoagulation.rst
@@ -109,47 +109,8 @@ In the case of catastrophic fragmentation, the largest mass of the fragment is t
 Collision Optimization
 ++++++++++++++++++++++
 `Zsom and Dullemond 2008 <https://www.aanda.org/articles/aa/abs/2008/38/aa09921-08/aa09921-08.html>`_ introduced a fine-tuning parameter, denoted as :math:`dm_{\rm{max}}`, into the algorithm to group collisions and thereby
-accelerate computation. It limits the maximum mass ratio for grouping collisions, and the collision rate :math:`C_{i,j}` is altered as follows,
- 
+accelerate computation. It limits the maximum mass ratio for grouping collisions by altering the collision rate of the particles.
 
-.. math:: 
- 
-    C^{\ast}_{i,j} = \frac{m{j}/m_{i}}{dm_{\rm{max}}}C_{i,j}
-
-This approach is particularly advantageous in regions such as dust traps, where collisions between particles with a broad size distribution
-occur frequently. Specifically, the probability of a collision between :math:`i`-th and :math:`j`-th can be reduced by a factor of, for example, 1000. When such a grouped collision occurs, 
-particle :math:`i`-th is assumed to accrete the equivalent of 1000 times the mass of particle j. While computationally efficient, this 
-simplification may not be realistic in dynamic systems, where particles can be advected across grid cells on timescales shorter than 
-those required to undergo such a large number of collisions. To address this, we propose an adaptive :math:`dm_{\rm{max}}`, which depends not 
-only on the mass ratio between colliding particles but also on the local collision rate and the spatial resolution of the grid.
-
-The adaptive value of :math:`dm_{\rm{max}}` for a collision between the :math:`i`-th and :math:`j`-th particles is computed as follows:
-
-.. math:: 
-
-    \tau_{i, \rm{transport}} = \tau^{\ast}_{i,j, \rm{collision}}
-
-.. math:: 
-    \frac{\Delta r}{v_{r, i}} = \frac{1}{C^{\ast}_{i,j}}
-
-.. math::
-
-    \frac{\Delta r}{v_{r, i}} = \frac{dm_{\rm{max}}}{m_{j}/m_{i}}\frac{1}{C_{i,j}}
-
-.. math:: 
-    
-    dm_{\rm{max}}=\frac{\Delta r}{v_{r, i}}\frac{m_{i}}{m_{j}}C_{i,j}
-
-
-where :math:`\Delta r` is the radial width of the grid cell. Here, the transport velocity :math:`v_{r, i}` accounts for all advection mechanisms except turbulence, which is excluded due to its stochastic nature.
-Hence, when the collision rate is high, :math:`dm_{\rm{max}}` can assume larger values, whereas in regions where transport is significant, :math:`m_{\rm{max}}` is reduced accordingly.
-Additionally, since transport in our simulations can occur in the vertical direction as well, we generalize our expression as follows:
-
-.. math:: 
-
-    dm_{\rm{max}}=min(\frac{\Delta r}{v_{r, i}}, \frac{\Delta z}{v_{z, i}}) \cdot \frac{m_{i}}{m_{j}}C_{i,j}
-
-Moreover, we constrain :math:`dm_{\rm{max}}` such that it does not introduce an error exceeding 1%. Consequently, :math:`dm_{\rm{max}}` is limited to a maximum value of 0.01. 
 
 Adaptive Grid
 +++++++++++++
