### 2025-10-08 48dc04b
features added for code comparison
