### 2025-12-04 #8619 @warrickball
Hi @vicky1997, I do apologise about the long delay. Your submission is still in the queue and waiting for an editor. I hesitate to estimate the further wait, since many reviews progressed only slowly over the summer and though I did say _at least_ 4 months, my similar estimates in other pre-reviews were also clearly too optimistic.

We've recently recruited a few more editors so hope to start progressing through the backlog but we can now see in our submission data that there has been a dramatic (~60%) and sustained increase in the number of submissions to JOSS since about March this year that we're struggling to keep up with.

This isn't the experience we want to offer authors but JOSS is run entirely by volunteers and we're doing everything we think we can to manage the submission rate without sacrificing quality.
