### 2026-01-27 #8619 @warrickball
Hi again @vicky1997, and sorry about the very long wait. We recently [announced changes to our requirements](https://blog.joss.theoj.org/2026/01/preparing-joss-for-a-generative-ai-future), which now appear in our docs on [what the paper should contain](https://joss.readthedocs.io/en/latest/paper.html#what-should-my-paper-contain), the [example paper](https://joss.readthedocs.io/en/latest/example_paper.html), and [scope definition](https://joss.readthedocs.io/en/latest/submitting.html#scope-and-significance). Could you update the paper to match the new format, include the newly-required sections? This may be by re-arranging existing material or writing new material. The maximum word count has increased from 1000 words to 1750 to accommodate the extra requirements. Once you have updated your paper, please comment here to let us know. At that point, we will assess your software against our updated scope criteria.

We want to be transparent that under these new criteria, some submissions that were previously in scope may now be deemed out of scope, and we may choose not to proceed with review. We recognise that this will be frustrating, especially for authors (like you) who have already been waiting many months for an editor to start handling a submission. The editorial board discussed this issue for some time and our consensus was that this choice will give JOSS's volunteer editors and reviewers the most clarity about what requirements to apply and support our commitment to ensuring JOSS publications recognise genuine scholarly contributions.
