### 2026-02-04 74c15ee
paper rewrite
