+++ paper/paper.bib
@@ -170,3 +170,92 @@ archivePrefix = {arXiv},
 	volume = 696,
 	pages = "A215",
 }
+
+@article{Andrews_2018,
+doi = {10.3847/2041-8213/aaf741},
+url = {https://doi.org/10.3847/2041-8213/aaf741},
+year = {2018},
+month = {dec},
+publisher = {The American Astronomical Society},
+volume = {869},
+number = {2},
+pages = {L41},
+author = {Andrews, Sean M. and Huang, Jane and Pérez, Laura M. and Isella, Andrea and Dullemond, Cornelis P. and Kurtovic, Nicolás T. and Guzmán, Viviana V. and Carpenter, John M. and Wilner, David J. and Zhang, Shangjia and Zhu, Zhaohuan and Birnstiel, Tilman and Bai, Xue-Ning and Benisty, Myriam and Hughes, A. Meredith and Öberg, Karin I. and Ricci, Luca},
+title = {The Disk Substructures at High Angular Resolution Project (DSHARP). I. Motivation, Sample, Calibration, and Overview},
+journal = {The Astrophysical Journal Letters},
+}
+
+@ARTICLE{Smoluchowski1916,
+       author = {{Smoluchowski}, M.~V.},
+        title = "{Drei Vortrage uber Diffusion, Brownsche Bewegung und Koagulation von Kolloidteilchen}",
+      journal = {Zeitschrift fur Physik},
+         year = 1916,
+        month = jan,
+       volume = {17},
+        pages = {557-585},
+       adsurl = {https://ui.adsabs.harvard.edu/abs/1916ZPhy...17..557S},
+      adsnote = {Provided by the SAO/NASA Astrophysics Data System}
+}
+
+@ARTICLE{Gillespie1975,
+       author = {{Gillespie}, Daniel T.},
+        title = "{An Exact Method for Numerically Simulating the Stochastic Coalescence Process in a Cloud.}",
+      journal = {Journal of the Atmospheric Sciences},
+         year = 1975,
+        month = oct,
+       volume = {32},
+       number = {10},
+        pages = {1977-1989},
+          doi = {10.1175/1520-0469(1975)032<1977:AEMFNS>2.0.CO;2},
+       adsurl = {https://ui.adsabs.harvard.edu/abs/1975JAtS...32.1977G},
+      adsnote = {Provided by the SAO/NASA Astrophysics Data System}
+}
+
+@article{Krijt2018,
+doi = {10.3847/1538-4357/aad69b},
+url = {https://doi.org/10.3847/1538-4357/aad69b},
+year = {2018},
+month = {aug},
+publisher = {The American Astronomical Society},
+volume = {864},
+number = {1},
+pages = {78},
+author = {Krijt, Sebastiaan and Schwarz, Kamber R. and Bergin, Edwin A. and Ciesla, Fred J.},
+title = {Transport of CO in Protoplanetary Disks: Consequences of Pebble Formation, Settling, and Radial Drift},
+journal = {The Astrophysical Journal},
[…]
