+++ paper/paper.md
@@ -26,23 +26,37 @@ bibliography: paper.bib
 
 # Summary
 
-`mcdust` is a parallel simulation code for dust evolution in protoplanetary disks. The code is written in `FORTRAN90` and parallelized with `OpenMP`. The code models dust collisional evolution and transport in the vertical and radial directions. The currently included collisional outcomes are dust growth by sticking, fragmentation of dust particles and erosion, where a small particle chips a portion of the large particle. We employ a representative particle approach detailed in @Zsom2008 to track a limited number of particles instead of tracking every particle, saving computational time. We have a static power-law gas disk with temperature assumed to be vertically isothermal. Dust coagulation depends on the local gas properties and therefore we bin particles into grids and perform collisions. We make use of an adaptive grid approach where we make sure that each cell has equal number of representative particles. This guarantees that there are always sufficient particles to resolve the physics of collisions. Figure below shows a sketch of our adaptive grid model and the different physical processes simulated by `mcdust`. The details of the physics of the code is explained in @Drazkowska2013.
-
-![](Figures/gridnew.png)
+The Solar System and other planetary systems formed in a disk of gas and dust around the young forming central star called a protoplanetary disk. The overall gas dynamics, the interaction between the gas and dust and the collisions between the dust particles leave an imprint on the formation pathways of the planets and therefore, on life. Recent observations with telescopes such as ALMA [@Andrews_2018] show a diversity of dust structures in protoplanetary disks and it is imperative that we have models that are able to describe the observations and more generally, to help move forward towards the goal of an end-to-end planet formation theory. We present `mcdust`, a parallel simulation code aimed at modelling the evolution of dust in protoplanetary disks to better understand the dust size distributions and dust collisional evolution.
 
 # Statement of Need
 
-Modelling dust coagulation and evolution is an essential part of understanding planet formation and protoplanetary disks. The evolution of micrometer-sized dust grains to millimeter-sized grains sets the tone for planetesimal formation. Dust evolution is inextricably linked with the formation of substructures in disks. Dust dynamics is also of prime importance when understanding protoplanetary disk chemistry. Therefore very important to understand in detail the processes involved in dust growth and dynamics and how they influence different aspects of disk dynamics and planet growth. 
+Modelling dust coagulation and evolution is an essential part of understanding protoplanetary disks and planet formation. The evolution of micrometer-sized dust grains to millimeter-sized aggregates sets the tone for planetesimal formation. Dust evolution is inextricably linked with the formation of substructures in disks. Dust dynamics is also of prime importance when understanding protoplanetary disk chemistry. Therefore it is very important to understand the processes involved in dust growth and dynamics and how they influence different aspects of disk dynamics and planet growth in detail.
 
-Dust coagulation is generally modelled with the Smoluchowski equation, an integro-differential equation. Some examples of open source codes that solve the Smoluchowski equation are `dustpy` [@Stammler2022], a 1D code to simulate gas and dust evolution and `cuDisc`[@Robinson2024], a 2D code modelling dust coagulation and disk evolution. Both of these codes are Eulerian in nature, i.e. they follow volume rather than mass. The drawback with these methods is that it is difficult to track histories of dust particles. Furthermore, adding a property to track adds a dimension to solving the Smoluchowski Equation [@Stammler2017]. Lagrangian methods can solve this issue partly by allowing us to track histories of particles. `PHANTOM` [@Price], is an open source Smoothed Particle Hydrodynamics (SPH) code that models models dust growth using a monodisperse growth or a 'single-size' approximation and do not solve the Smoluchowski equation for dust growth [@Vericel2021]. 
[…]
