### 2026-02-04 e8dab1d
paper rewrite
