+++ paper/paper.bib
@@ -258,4 +258,26 @@ abstract = {It is believed that satellites of giant planets form in circumplanet
 }
 
 
+@article{Birnstiel2012,
+	author = {{Birnstiel, T.} and {Klahr, H.} and {Ercolano, B.}},
+	title = {A simple model for the evolution of the dust population  in protoplanetary disks},
+	DOI= "10.1051/0004-6361/201118136",
+	url= "https://doi.org/10.1051/0004-6361/201118136",
+	journal = {Astronomy & Astrophysics},
+	year = 2012,
+	volume = 539,
+	pages = "A148",
+	month = "",
+}
+
+@article{Pfeil2024,
+	author = {{Pfeil, Thomas} and {Birnstiel, Til} and {Klahr, Hubert}},
+	title = {TriPoD: Tri-Population size distributions for Dust evolution - Coagulation in vertically integrated hydrodynamic simulations of protoplanetary disks},
+	DOI= "10.1051/0004-6361/202449337",
+	url= "https://doi.org/10.1051/0004-6361/202449337",
+	journal = {Astronomy & Astrophysics},
+	year = 2024,
+	volume = 691,
+	pages = "A45",
+}
 
