+++ paper/paper.md
@@ -32,15 +32,15 @@ The Solar System and other planetary systems formed in a disk of gas and dust ar
 
 Modelling dust coagulation and evolution is an essential part of understanding protoplanetary disks and planet formation. The evolution of micrometer-sized dust grains to millimeter-sized aggregates sets the tone for planetesimal formation. Dust evolution is inextricably linked with the formation of substructures in disks. Dust dynamics is also of prime importance when understanding protoplanetary disk chemistry. Therefore it is very important to understand the processes involved in dust growth and dynamics and how they influence different aspects of disk dynamics and planet growth in detail.
 
-Dust coagulation is generally modelled with the Smoluchowski equation[@Smoluchowski1916], an integro-differential equation. The drawback with the method is that it is difficult to track histories of dust particles. Furthermore, adding a property to track adds a dimension to solving the Smoluchowski Equation [@Stammler2017].
+Dust coagulation is generally modelled with the Smoluchowski equation [@Smoluchowski1916], an integro-differential equation. The drawback with the method is that it is difficult to track histories of dust particles. Furthermore, adding a property to track adds a dimension to solving the Smoluchowski Equation [@Stammler2017].
 
-Alternatively one can model dust coagulation by performing Monte Carlo simulations. Monte Carlo methods are well suited for stochastic processes and that makes it a suitable method for dust coagulation[@Gillespie1975]. For such a method, the representative particle approach [@Zsom2008] is very suitable and the representative particle approach is Lagrangian in nature, meaning we track the particles and hence their histories. We can add properties to the particles that can be tracked without much computational complexity. This advantage is very useful when we want to combine dust evolution and chemistry, where we want to look at the thermal histories of particles and their compositions when chemical models are included. 
+Alternatively one can model dust coagulation by performing Monte Carlo simulations. Monte Carlo methods are well suited for stochastic processes and that makes it a suitable method for dust coagulation [@Gillespie1975]. For such a method, the representative particle approach [@Zsom2008] is very suitable and the representative particle approach is Lagrangian in nature, meaning we track the particles and hence their histories. We can add properties to the particles that can be tracked without much computational complexity. This advantage is very useful when we want to combine dust evolution and chemistry, where we want to look at the thermal histories of particles and their compositions when chemical models are included. 
 
-Although there have been several publications that have made use of the Monte Carlo method to model dust coagulation in protoplanetary disks [@Krijt2018, @Houge2023], there is no available open source code that models dust coagulation using Monte Carlo methods and such a code would be very helpful as an alternative way for modelling dust coagulation.
+Although there have been several publications that have made use of the Monte Carlo method to model dust coagulation in protoplanetary disks [@Krijt2018;@Houge2023], there is no available open source code that models dust coagulation using Monte Carlo methods and such a code would be very helpful as an alternative way for modelling dust coagulation.
 
 # State of the field
 
[…]
