### 2026-02-04 a50014d
paper rewrite
