+++ paper/paper.bib
@@ -259,7 +259,7 @@ abstract = {It is believed that satellites of giant planets form in circumplanet
 
 
 @article{Birnstiel2012,
-	author = {{Birnstiel, T.} and {Klahr, H.} and {Ercolano, B.}},
+	author = {{Birnstiel}, T. and {Klahr, H.} and {Ercolano, B.}},
 	title = {A simple model for the evolution of the dust population  in protoplanetary disks},
 	DOI= "10.1051/0004-6361/201118136",
 	url= "https://doi.org/10.1051/0004-6361/201118136",
@@ -271,7 +271,7 @@ abstract = {It is believed that satellites of giant planets form in circumplanet
 }
 
 @article{Pfeil2024,
-	author = {{Pfeil, Thomas} and {Birnstiel, Til} and {Klahr, Hubert}},
+	author = {{Pfeil}, Thomas and {Birnstiel, Til} and {Klahr, Hubert}},
 	title = {TriPoD: Tri-Population size distributions for Dust evolution - Coagulation in vertically integrated hydrodynamic simulations of protoplanetary disks},
 	DOI= "10.1051/0004-6361/202449337",
 	url= "https://doi.org/10.1051/0004-6361/202449337",
