### 2026-02-04 #8619 @vicky1997
Hi @warrickball!

Thank you very much for letting us know about the change and we understand the need for the change in the guidelines. We have rewritten the paper to fit the new guidelines and updated it in the paper/ directory of the repository. Thank you very much.

Best,
Vignesh
