### 2026-02-05 #8619 @warrickball
Hi @vicky1997, and thanks for these updates. I believe the article remains in scope but two minor things stand out.

1. It wasn't clear to me from the text whether this submission includes the original code from Drążkowska et al. (2013) or whether this is a fresh implementation of the same questions. If it does include the original code, then I think the *State of the Field* can mention that mcdust was created before many of the other codes were published. If it does not, then I don't think it's fair to refer in the *Research Impact Statement* to papers that used the old code to demonstrate the impact of the new one. It does demonstrate that one can reasonably expect the new code to have a research impact.

2. It's relatively minor but note that the images don't have captions and aren't linked to from the text. Have a look at our docs on [figures and images](https://joss.readthedocs.io/en/latest/paper.html#figures-and-images) for the appropriate way of including figures. Though they aren't rendered as floats like mainstream journal articles, they should be treated as such because I might want to move them around (without changing the order) if they otherwise leave lots of empty space in the rendered paper.
