### 2026-02-06 d0b8a23
update npar after restart
