### 2026-02-06 2bf5f4a
paper figure caption added
