### 2026-02-06 d152adb
paper edits
