+++ paper/paper.md
@@ -40,11 +40,11 @@ Although there have been several publications that have made use of the Monte Ca
 
 # State of the field
 
-Several codes exist to model dust growth in protoplanetary disks and we focus on discussing the open sources codes in the field. There are semi-analytical methods like `twopoppy` [@Birnstiel2012] and `TriPoD` [@Pfeil2024] that model the overall dust size distributions without modelling full dust coagulation. Some examples of open source codes that model full dust coagulation are `dustpy` [@Stammler2022], a 1D code to simulate gas and dust evolution and `cuDisc` [@Robinson2024], a 2D (radial-vertical) code modelling dust coagulation and disk evolution. Both of these codes solve the Smoluchowski equation are Eulerian in nature, i.e. they follow volume rather than mass. `PHANTOM` [@Price], is an open source Smoothed Particle Hydrodynamics (SPH) code that models models dust growth using a monodisperse growth or a 'single-size' approximation and do not solve the Smoluchowski equation for dust growth [@Vericel2021].
+Several codes exist to model dust growth in protoplanetary disks and we focus on discussing the open sources codes in the field. There are semi-analytical methods like `twopoppy` [@Birnstiel2012] and `TriPoD` [@Pfeil2024] that model the overall dust size distributions without modelling full dust coagulation. Some examples of open source codes that model full dust coagulation are `dustpy` [@Stammler2022], a 1D code to simulate gas and dust evolution and `cuDisc` [@Robinson2024], a 2D (radial-vertical) code modelling dust coagulation and disk evolution. Both of these codes solve the Smoluchowski equation are Eulerian in nature, i.e. they follow volume rather than mass. `PHANTOM` [@Price], is an open source Smoothed Particle Hydrodynamics (SPH) code that models models dust growth using a monodisperse growth or a 'single-size' approximation and do not solve the Smoluchowski equation for dust growth [@Vericel2021]. 
 
-`mcdust` models dust coagulation and transport in the vertical and radial directions. The currently included collisional outcomes are dust growth by sticking, fragmentation of dust particles and erosion, where a small particle chips a portion of the large particle. We employ a representative particle approach detailed in @Zsom2008 to track a limited number of particles instead of tracking every particle, saving computational time. We have a static power-law gas disk with temperature assumed to be vertically isothermal. Figure \ref{fig:grid} shows the overview of the different processes in `mcdust` and a representation of the adaptive grid method.
+`mcdust` models dust coagulation and transport in the vertical and radial directions. The currently included collisional outcomes are dust growth by sticking, fragmentation of dust particles and erosion, where a small particle chips a portion of the large particle. We employ a representative particle approach detailed in @Zsom2008 to track a limited number of particles instead of tracking every particle, saving computational time. We have a static power-law gas disk with temperature assumed to be vertically isothermal. Dust coagulation depends on the local gas properties and therefore we bin particles into grids in order to perform collisions. We make use of an adaptive grid approach where we make sure that each cell has equal number of representative particles. This guarantees that there are always sufficient particles to resolve the physics of collisions. Figure \ref{fig:grid} shows a sketch of our adaptive grid model and the different physical processes simulated by `mcdust`. `mcdust` in its first iteration was introduced in @Drazkowska2013 and we refer the reader to @Drazkowska2013 for the details of the physics of the code.
 
[…]
