### 2026-02-06 fffa01a
paper edits
