### 2026-02-06 #8619 @vicky1997
Hi @warrickball ,

Thank you very much for your reply.

1.  The submission includes the original code from Drążkowska et al. (2013), which has been optimised to make it faster, and some software design changes have been made to make it more usable and adaptable. So the basic code remains the same that has been used over the years. I have added this information in the _State of the Field_ section. Thank you for the comment.
2. I have added captions and have referenced them internally in the text as well. Hope this is okay!

Hope I have answered your comments. I have updated the paper with the said changes. Thank you very much!

Best,
Vignesh
