### 2026-02-07 #8619 @warrickball
Thanks for those updates. I think yes, the submission is once again ready for review and in effect return to the submission queue. I realise and apologise that you've already been waiting so long but at the moment I expect it'll still be a few more months before an editor becomes available. Sorry!
