### 2026-02-13 a3be685
raddrift flag removed from collisions
Removed conditional compilation for radial drift velocity calculations.
