### 2026-02-13 8def732
raddrift flag removed from collisions
Removed RADRIFT conditional compilation for velocity drift calculations.
