### 2026-02-16 d5bfb2d
Update attribute name for mass_of_swarm in HDF5
