### 2026-02-17 61599b1
Fix preprocessor directive placement in advection.F90
