### 2026-03-03 #10140 @editorialbot
[REVIEW]: mcdust: A 2D Monte Carlo code for dust coagulation in protoplanetary disks
**Submitting author:** @vicky1997 (<a href="http://orcid.org/0000-0002-2451-9574">Vignesh Vaikundaraman</a>)
**Repository:** https://github.com/vicky1997/mcdust
**Branch with paper.md** (empty if default branch): master
**Version:** v1.0
