### 2026-03-24 #10140 @AdrienOri
Hey @vicky1997 :) I am going through the checklist for what needs to be included in the paper. The Sections 'Software Design', 'Research Impact Statement' and 'AI usage disclosure' all seem to be required sections but I don't see them in the manuscript. I'd imagine the Software Design part to be mainly discussed in previous papers by Joanna (2014, 2019). Don't you think it'd be worth piling things up in this new paper?
