### 2026-03-24 #10140 @vicky1997
Hey @AdrienOri !

Thank you for your reply and taking the time to review the paper and the code. I can see the sections 'Software Design', 'Research Impact Statement' and 'AI usage disclosure' in the article proof version few replies above. Please do let me know if I am looking at something else.

Best,
Vignesh
