### 2026-03-25 #10140 @AdrienOri
Good morning @vicky1997, congrats for the co-authored paper of Linn I saw today! I had a couple of first runs with mcdust yesterday and I wanted to give you a few feedbacks on the mcdust docs and some problems I encountered.
Though the issues I got are really minor, I think it's important for potential new mcdust users to quickly be able to have a successful first run and easy data analysis, so they are motivated to pursue their exploration :)

(1) Upon installation, I used `git clone git@github.com:vicky1997/mcdust.git` as given in the docs, but didn't have the rights to download it. However, the following worked `git clone https://github.com/vicky1997/mcdust.git`

(2) Fortran-OpenMP is mentioned in the Github docs as a prerequisite but not in the official Github

(3) in the docs, First Run - point (1.), please mention that the `make` is to be run in the main mcdust directory

(4) in the docs, First Run - point (3.), ` ./default /setups/default/setup.par` fails, shouldn't it be replaced by ` ./default setups/default/setup.par` instead

(5) in the docs, First Run: Once compiling my first run from my terminal, it was a bit difficult to understand what was happening and how my simulation was progressing. It may help the new user if you specifically mentioned that the time evolution of the simulation can be found in timesout.dat. You could also mention it more explicitely in the Input and Output section.

(6) in the docs, Restart: Please add the `restart` parameter to setup.par in setups/default. For a new user passing from the First Run to Restart as a test, it may be confusing that `restart` isn’t in the setup.par of setups/default

(7) More generally, introducing what the default run is doing may help the new users understanding the physics behind their first compilation? e.g., “To ensure of the proper installation of mcdust, we invite the reader to attempt a compilation of a test simulation located in setup/default. It consists in a disc with ………". You could describe the dimensional setup (1D or 2D), the disc mass etc.

(7) mcdust_examples.ipynb: fantastic!! Perhaps also add the description about the input parameter ‘time’ in the Data Visualisation part. The same way ‘step’ is currently describe. So it is clear to the reader that ‘time’ is the actual time in years, not the index of the snapshot file (except for the last one snapshot, input as -1).

(8) in mcdust_examples.ipynb: I tried plotting Plots.sigma_d but sigma_d was setup to none and it failed. Opening the mcdust.py file, I found out about the need to run calculate_properties() first, but I had an error as well. I found out it's because the parameter `rdis` is in [au] while the others are in cm, so the histogram returns an array of 0s. Converting rdis to cm units does the job.
The line `rcounts[:-1] *self.mswarm / (2 * np.pi * rcents*drdis)` also returns an error because rcounts doesn't have the right dimension. What's the purpose of [:-1] after rcounts?

These are all minor comments, but I believe they are necessary to convert more easily 'potential curious users' into 'actual users' of mcdust. Great job otherwise!! Let me know if you have any questions.
