### 2026-03-25 #10140 @AdrienOri
Some follow-up questions:

(1) Upon varying `data_directory` in `setup.par` to something else than "data" (e.g., "data_test"), it seems that it doesn't consider my input but still ends up writing in "data".

(2) I'd like to test the 1D vertical column setup in `preprocs.opt`. Is it simply a matter of removing the # to `#CFLAGS+=-D VERTICALMODEL`? It'd be worth having a setup run in /setups for a 1D vertical column to test that case and compare with your 2025 paper :)

(3) It's not clear how the gas is treated in mcdust. In Joanna's 2013 paper I can read that it is not evolved, such that I'd imagine gas to not be flowing. But at the same time, the gas advection velocity is considered in the dust velocity (Eq.6). With the 2D (r-z) structure, are there strong advection flows in the disc atmosphere?

(4) About the adaptative grid algorithm: Is there a limit to the volume a cell can reach? The fundamental assumption in the MC approach is that all particles in a given cell should be able to collide with one another in a given timestep. If the cell gets too large (let's say 5 AU wide) and contains only a few representative particles moving at low velocities, doesn't it break the main assumption of MC?
