### 2026-04-16 #10140 @mlombart
Hello @vicky1997,
I apologize for the delay. I would like to congratulate you for this nice work on developing the code mcdust. This new tool is very important for the astrophysics community.
I have read the paper and tried some runs with the code. I give you my feedbacks about some minor issues I encountered and also some suggestions for future development. I will first give you feedbacks on the paper and then about the code and documentation.

_Comments for the paper:_
The manuscript is well written and I give some comments in the following.

**In the summary part:**
On line 14, I suggest to add that the code mcdust is based on Monte Carlo method, to directly show to the reader which numerical method is used.

**In the part “A statement of need”:**
(1) On line 26, can you explain what is the meaning of “to track histories of dust particles” ? If I well understood, you mentioned the track of positions of dust particles. In that case, I do not understand why tracing the history of dust particles is a drawback of the Smoluchowski equation. In this equation, there is no dust particle advection. Therefore, the history of dust position would be given by the hydrodynamics code in which the Smoluchowski equation is solved.
(2) On line 28, other references for the multidimensional Smoluchowski equation can be added such as [Blum, 2006](https://doi.org/10.1080/00018730601095039) and [Okuzumi et al., 2009](https://doi.org/10.1088/0004-637X/707/2/1247)

**In the part “State of field”:**
(1) On line 49, I guess it is “solving” instead of “solve …”.
(2) On line 51, there are two words “models”, one can be removed.
(3) On line 59, can you explain the assumption to have a static gas and a dynamic dust in order to have a clear view of the gas and dust modeling for the reader ?
(4) On line 74, the post-processing data from hydrodynamic simulations is very interesting, very useful for the community. If it is possible, I would suggest to give more details in the paper on how the post-processing works and in the documentation to guide reader to use mcdust to post-process hydrodynamics simulations.

**In the part “Benchmark”:**
(1) Since mcdust is based on the code from Drążkowska et al., 2013, I would recommend to add all the tests presented in this paper, such as dust coagulation tests. A test for the dust fragmentation would be very interesting to add, with detailed about the analytical solution detailed in [Lombart et al., 2024](https://ui.adsabs.harvard.edu/abs/2024MNRAS.533.4410L)
(2) On line 112, having no issue with the artificially sped-up growth is an important statement and has to be shown numerically by testing mcdust with the analytical solution for the Smoluchowski equation.
