### 2026-04-16 #10140 @mlombart
Here I put my feedback on the code and documentation.

The code and documentation are well written, almost all the functions are commented which makes the code easy to read and understand.

(1) As mentioned by @AdrienOri, I did not have issue about the installation with the git clone, it worked directly on my side.

(2) In the README, I would suggest to add details on prerequisites as presented in the documentation.

(3) I confirm the points mentioned by @AdrienOri, I faced the same issue with the first run.
The command should be `./default setups/default/setup.par`, similar to the test command.

(4) I also have the issue with changing name for data directory in file `setup.par`. The new directory name is not taking into account. The outputs of the simulation are still saved in `data` directory.

(5) I had an issue with the restart steps. Some details are missing. It is not mentioned that the flag in `preprocs.opt` has to be activated for RESTART. Then, which command should be run ? If I run the usual command `./default setups/default/setup.par`, I encounter errors with hdf5. Please add more details for the reader to successfully do a restart run.

(6) I tried to run the default simulation in debug mode, but I obtain this error `Program received signal SIGILL: Illegal instruction.` Is it expected ?

(7) On my MacBook Pro with 4 CPUs, the default run took around 1h to finish. Is it the expect elapsing time for this simulation ?

(8) In the documentation at dust coagulation, as I mentioned, I would suggest to add the tests presented in Drążkowska et al., 2013. And if possible, I would recommend to add the test for dust fragmentation detailed in [Lombart et al., 2024](https://ui.adsabs.harvard.edu/abs/2024MNRAS.533.4410L).

(9) In the documentation at the part visualization, the jupyter notebook is great with all needed details ! Thank you for this work. It is just missing the pre-requisites for python module to install.

(10) As mentioned for the paper, it would be very interesting to have details or example on how to use mcdust for post-processing output from hydrodynamics simulations.

(11) In file `script/mcdisk.py`, on line 211, it is missing the description of the function `vel_vn`.

(12) In file `script/mcdust.py`, on line 47, it is missing the description of the function `read_hdf5`.

(13) In file `collisions.f90`:
        - on line 175, it is better to write from which Til Birnstiel's paper the implementation is coming from.
        - on line 354, I suggest to give some details (example/template) for the user to use their own collision model.

(14) In file `discstruct.f90`, on line 1, maybe it would be interesting to give an example or template to help reader to write their own disc structure.

_Suggestion for future developments:_
In order to improve the code mcdust, I would suggest to implement, if it is possible, the improvement of the representative particle scheme detailed in [Beutel et al., 2023](https://www.aanda.org/articles/aa/abs/2023/02/aa44955-22/aa44955-22.html) and [Beutel et al., 2024](https://www.aanda.org/articles/aa/abs/2024/04/aa47304-23/aa47304-23.html).

Let me know if you have questions on my feedbacks. Would you like I open issues in the git repository for all the minors issues I encountered or suggestions ?

Congratulations again ! It is a very nice work !
