### 2026-04-16 42554f6
default run desc + 1D example added
