+++ Readme.md
@@ -4,19 +4,19 @@
 
 ## Prerequisites
 
-`gfortran` `hdf5-serial` `python`
+`gfortran` `hdf5-serial` `python` `OpenMP`
 
 To install the required software in Ubuntu(this requires root permissions):
 `sudo apt-get install gfortran`
 `sudo apt-get install libhdf5-serial-dev`
 
 Python is not required to run the code. But if you want to use the routines to read/write data from the simulation you will need a python installation.
-
+v 
 ## To compile the code: 
 
 `make SETUP_FILE=default`
 
-This uses the setup files from the default run in `/setups/default/` 
+This uses the setup files from the default run in `setups/default/` 
 The setup files can be modified to suit your needs and can be added to the `setups` directory. An executable in the name of the setup file will be created in the root directory which is `default` in this case.
 
 parameter file: setups/default/setup.par
@@ -28,7 +28,7 @@ To run the unit tests:
 
 `export OMP_NUM_THREADS=number_of_CPUs`
 
-`./default /setups/default/setup.par`
+`./default setups/default/setup.par`
 
 To clear the run
 
