### 2026-05-28 d5a4882
1D model docs
