### 2026-05-28 c9f4223
1D model docs
