+++ docs/source/dataviz.rst
@@ -1,2 +1,0 @@
-Data analysis and visualization using Python
-============================================
