+++ docs/source/example_2.rst
@@ -1,4 +1,4 @@
-Example 2: Adding dust growth outcomes
+Example 3: Adding dust growth outcomes
 ===========================
 
 Coming Soon..
