+++ docs/source/example_1.rst
@@ -1,4 +1,4 @@
-Example 1: Tracking histories
+Example 2: Tracking histories
 ====================
 
 Coming Soon..
