+++ docs/source/example_1.rst
@@ -1,4 +1,4 @@
-Example 2: Tracking histories
+Example: Tracking histories
 ====================
 
 Coming Soon..
