+++ docs/source/index.rst
@@ -19,8 +19,8 @@
     standardmodel
     dustdynamics
     dustcoagulation
-    notebooks/mcdust_example.ipynb
-    notebooks/verticalcol.ipynb
+    notebooks/mcdust_example
+    notebooks/verticalcol
     example_1
     example_2
     citation
