+++ docs/source/index.rst
@@ -5,7 +5,7 @@
 
 | ``mcdust`` is a 2D (r-z) code that simulates the vertical and radial evolution and growth of dust. The code makes use of the representative particle approach to simulate the dust particles and uses Monte Carlo methods to simulate dust growth.
 
-| The details are ``mcdust`` are presented in  `Drążkowska et al 2013 <https://www.aanda.org/articles/aa/abs/2013/08/aa21566-13/aa21566-13.html>`_ & `Vaikundaraman et al. (in review in JOSS)<https://arxiv.org/abs/2507.21239>`_. The code units are in cgs and we use a cylindrical coordinate system (r-z) in the code.
+| The details are ``mcdust`` are presented in  `Drążkowska et al 2013 <https://www.aanda.org/articles/aa/abs/2013/08/aa21566-13/aa21566-13.html>`_ & `Vaikundaraman et al. (in review in JOSS) <https://arxiv.org/abs/2507.21239>`_ . The code units are in cgs and we use a cylindrical coordinate system (r-z) in the code.
 
 | Here we outline the basic setup and usage of the code and some ways in which one can make use of the features of the code.
 
@@ -19,7 +19,7 @@
     standardmodel
     dustdynamics
     dustcoagulation
-    notebooks/verticalcol.ipynb
+    notebooks/mcdust_example.ipynb
     example_1
     example_2
     citation
