+++ docs/source/conf.py
@@ -20,6 +20,10 @@ extensions = [
     'nbsphinx'
 ]
 
+exclude_patterns= [
+    '**.ipynb_checkpoints',
+]
+
 intersphinx_mapping = {
     'python': ('https://docs.python.org/3/', None),
     'sphinx': ('https://www.sphinx-doc.org/en/master/', None),
