+++ docs/requirements.txt
@@ -1,3 +1,6 @@
-sphinx==7.1.2
-sphinx_rtd_theme==1.3.0rc1
+.
+sphinx
+sphinx_rtd_theme
 nbsphinx
+sphinx-automodapi
+pandoc
