+++ docs/requirements.txt
@@ -1,6 +1,4 @@
-.
 sphinx
 sphinx_rtd_theme
 nbsphinx
 sphinx-automodapi
-pandoc
