+++ docs/source/index.rst
@@ -20,7 +20,6 @@
     dustdynamics
     dustcoagulation
     notebooks/mcdust_example.ipynb
-    verticalcol.ipynb
     example_1
     citation
     
