+++ docs/requirements.txt
@@ -1,5 +1,4 @@
 sphinx
 sphinx_rtd_theme
-nbsphinx
 sphinx-automodapi
-myst_nb
+nbsphinx
