+++ docs/source/conf.py
@@ -17,7 +17,7 @@ extensions = [
     'sphinx.ext.autodoc',
     'sphinx.ext.autosummary',
     'sphinx.ext.intersphinx',
-    "nbsphinx",
+    'nbsphinx',
 ]
 
 exclude_patterns= [
