+++ docs/requirements.txt
@@ -2,4 +2,3 @@ sphinx
 sphinx_rtd_theme
 sphinx-automodapi
 nbsphinx
-pandoc
