### 2026-06-03 2232fab
kernel tests added
