+++ paper/paper.bib
@@ -71,23 +71,7 @@ archivePrefix = {arXiv},
       adsnote = {Provided by the SAO/NASA Astrophysics Data System}
 }
 
-@ARTICLE{2007A&A...461..215O,
-       author = {{Ormel}, C.~W. and {Spaans}, M. and {Tielens}, A.~G.~G.~M.},
-        title = "{Dust coagulation in protoplanetary disks: porosity matters}",
-      journal = {Astronomy & Astrophysics},
-     keywords = {ISM: dust, extinction, planetary systems: formation, planetary systems: protoplanetary disks, accretion, accretion disks, Astrophysics},
-         year = 2007,
-        month = jan,
-       volume = {461},
-       number = {1},
-        pages = {215-232},
-          doi = {10.1051/0004-6361:20065949},
-archivePrefix = {arXiv},
-       eprint = {astro-ph/0610030},
- primaryClass = {astro-ph},
-       adsurl = {https://ui.adsabs.harvard.edu/abs/2007A&A...461..215O},
-      adsnote = {Provided by the SAO/NASA Astrophysics Data System}
-}
+
 
 @article{Vericel2021,
     author = {Vericel, Arnaud and Gonzalez, Jean-François and Price, Daniel J and Laibe, Guillaume and Pinte, Christophe},
@@ -281,3 +265,117 @@ abstract = {It is believed that satellites of giant planets form in circumplanet
 	pages = "A45",
 }
 
+@article{Blum2006,
+author = {Jürgen Blum},
+title = {Dust agglomeration},
+journal = {Advances in Physics},
+volume = {55},
+number = {7-8},
+pages = {881--947},
+year = {2006},
+publisher = {Taylor \& Francis},
+doi = {10.1080/00018730601095039},
+
+
+URL = { 
+    
+        https://doi.org/10.1080/00018730601095039
+    
+    
+
+},
+eprint = { 
+    
+        https://doi.org/10.1080/00018730601095039
+    
+    
+
+}
+
+}
+
+@article{Okuzumi_2009, 
+doi = {10.1088/0004-637X/707/2/1247}, 
+url = {https://doi.org/10.1088/0004-637X/707/2/1247}, 
+year = {2009}, 
+month = {dec}, 
+publisher = {The American Astronomical Society}, 
+volume = {707}, 
+number = {2}, 
+pages = {1247}, 
+author = {Okuzumi, Satoshi and Tanaka, Hidekazu and Sakagami, Masa-aki}, 
+title = {NUMERICAL MODELING OF THE COAGULATION AND POROSITY EVOLUTION OF DUST AGGREGATES}, 
+journal = {The Astrophysical Journal},  
+} 
+
+@article{Gurrutxaga_2026,
+doi = {10.3847/1538-4357/ae6104},
+url = {https://doi.org/10.3847/1538-4357/ae6104},
+year = {2026},
+month = {may},
+publisher = {The American Astronomical Society},
+volume = {1003},
+number = {2},
+pages = {132},
+author = {Gurrutxaga, Nerea and Drążkowska, Joanna and Vaikundaraman, Vignesh and Kleine, Thorsten},
+title = {Carbonaceous Chondrites Provide Evidence for Late-stage Planetesimal Formation in a Pressure Bump},
+journal = {The Astrophysical Journal},
+}
+
+@article{ Houge2025,
[…]
