+++ unit_tests/kerneltest.F90
@@ -1,0 +1,265 @@
+program testkernel
+
+    use types
+    use hdf5output
+    use hdf5
+    use parameters, only: path, datadir, read_parameters
+
+    implicit none
+    !  ------- set parameters here --------------------------------------------------------------------------------------
+   integer, parameter             :: ntot = 1000      ! total number of representative particles (swarms)
+   real, parameter                :: dens = 1.       ! dust volume mass density
+   real, parameter                :: mswrm = 10.e20  ! mass of the swarm
+#ifdef FRAGTEST
+   real, parameter                :: m0 = 1e-6
+   real, parameter                :: gamma = 1e4
+#else
+   real, parameter                :: m0 = 1.         ! mass of the monomer (initial mass of the particles)
+#endif
+   real, parameter                :: tend = 20.      ! maximum time of the simulation
+   real, parameter                :: dtout = 4.      ! time step for the output
+   integer, parameter             :: nbins = 250     ! number of bins for output histograms
+   integer, parameter             :: repeat = 10     ! number of times you want to repeat a simulation
+   real, parameter                :: dmmax = 0.001
+   ! -------------------------------------------------------------------------------------------------------------------
+
+   type(swarm), dimension(ntot)      :: swarms             ! list of the swarms
+   integer                        :: nri               ! index of representative partcile to undergo the next collision
+   integer                        :: nrk               ! index of physical partcile to undergo the next collision
+   real                           :: vol               ! volume of the cell
+   real                           :: time = 0.0        ! time
+   real                           :: dt             ! time step between the subsequent collisions
+   real                           :: tout = dtout      ! time of the next output
+   real, dimension(ntot,ntot)     :: colrates          ! matrix of the collision rates between the particles
+   real, dimension(ntot)          :: colrates_rp       ! probability of choosing given representative particles
+   real                           :: totrate           ! total collision rate
+   real                           :: rand              ! random number
+   real                           :: fin               ! temporary value
+   character(len=11)              :: fname             ! output file name
+   real, dimension(ntot)          :: colrates_old_col  ! for optimization reasons
+   real, dimension(ntot)          :: accelncol ! collision optimisation
+   integer                        :: i, j, cont_time, loops, kk
+   character(len=100)               :: loops_str
+   character(len=100)               :: command
+   character(len=100)               :: ctrl_file
+   character(len=200)               :: filepath
+   type(hdf5_file_t)                :: file
+   real, dimension(:), allocatable :: t_arr
+   real, dimension(nbins)         :: m2fm              ! mass distribution function
+   real, dimension(nbins+1)       :: mgrid             ! mass grid
+   integer                        :: k = 1
+   real                           :: ll, lll, nord
+
+  datadir = 'kernels'
+  call get_command_argument(1, ctrl_file)
+  write(*,*) 'Reading parameters...'
+  call read_parameters(ctrl_file)
+
+#ifdef LINEARKERNEL
+   !dtime = 20.
+   allocate(t_arr(5))
+   t_arr = (/4., 8., 12., 16., 20./)
+#elif PRODUCTKERNEL
[…]

[…]
