### 2026-06-03 b71a21d
kernel docs
