+++ docs/requirements.txt
@@ -2,3 +2,4 @@ sphinx
 sphinx_rtd_theme
 sphinx-automodapi
 nbsphinx
+pandoc
