### 2026-06-03 e3e8419
paper update
