+++ paper/paper.md
@@ -76,7 +76,7 @@ We ran a simulation for 10000 years. Figure \ref{fig:benchmark} shows the dust s
 ![A comparison of the dust surface density $\sigma_d$ as a function of particle mass and distance from star between `mcdust` (left) and `dustpy` (right).\label{fig:benchmark}](Figures/dustpyvsmcdust.png){width=75%}
 
 
-It is evident from Figure \ref{fig:benchmark} that both the codes have similar overall outcomes but they do have certain differences. The most striking one is that `mcdust` does not provide coverage of the regions of parameter space that do not include sufficiently high fraction of the dust mass. This holds true for all Monte Carlo based codes. But with higher resolution simulation this issue can be overcome. The other important factor is that `mcdust` does not face the issue of artificially sped-up growth[@Drazkowska2014] that `dustpy` and other Smoluchowski equation based codes tend to encounter[@Stammler2022]. And the 2D r-z structure of `mcdust` also helps us to investigate processes like sedimentation driven coagulation  [@Drazkowska2013] that are not usually seen in 1D simulations like `dustpy`. This can be seen in image at around 50 AU where `mcdust` has larger surface densities higher masses when compared to `dustpy`. For a more detailed discussion of the differences between the two approaches to dust growth, i.e., the Monte Carlo method and the Smoluchowski equation approach we point the reader to @Drazkowska2014. Testing against analytical coagulation kernels are an important way to validate the coagulation models and we refer the reader to @Drazkowska2013 and the `mcdust` documentation for the different coagulation kernel tests.
+It is evident from Figure \ref{fig:benchmark} that both the codes have similar overall outcomes but they do have certain differences. The most striking one is that `mcdust` does not provide coverage of the regions of parameter space that do not include sufficiently high fraction of the dust mass. This holds true for all Monte Carlo based codes. But with higher resolution simulation this issue can be overcome. The other important factor is that `mcdust` does not face the issue of artificially sped-up growth[@Drazkowska2014] that `dustpy` and other Smoluchowski equation based codes tend to encounter[@Stammler2022]. And the 2D r-z structure of `mcdust` also helps us to investigate processes like sedimentation driven coagulation  [@Drazkowska2013] that are not usually seen in 1D simulations like `dustpy`. This can be seen in image at around 50 AU where `mcdust` has larger surface densities higher masses when compared to `dustpy`. For a more detailed discussion of the differences between the two approaches to dust growth, i.e., the Monte Carlo method and the Smoluchowski equation approach we point the reader to @Drazkowska2014. Benchmarking the simulations against analytical coagulation kernels are an important way to validate the coagulation models and we refer the reader to @Drazkowska2013 and the `mcdust` documentation for the different coagulation kernel tests.
 
 # AI usage disclosure
 No AI tools including generative AI tools have been used in the development of the software, creation of the documentation or in the writing of the paper. 
