+++ unit_tests/util_functions.F90
@@ -1,0 +1,216 @@
+module util_functions
+    implicit none
+
+    public :: lambert_w
+
+    contains
+
+    function lambert_w ( x, nb, l )
+
+        !*****************************************************************************80
+        !
+        !! lambert_w() approximates the Lambert W function.
+        !
+        !  Discussion:
+        !
+        !    The range requirement for the upper branch is:
+        !      -exp(-1) <= X.
+        !    The range requirement for the lower branch is:
+        !      -exp(-1) < X < 0.
+        !
+        !  Licensing:
+        !
+        !    This code is distributed under the MIT license.
+        !
+        !  Modified:
+        !
+        !    20 June 2023
+        !
+        !  Author:
+        !
+        !    Original Fortran77 version by Andrew Barry, S. J. Barry, 
+        !    Patricia Culligan-Hensley.
+        !    This version by John Burkardt.
+        !
+        !  Reference:
+        !
+        !    Andrew Barry, S. J. Barry, Patricia Culligan-Hensley,
+        !    Algorithm 743: lambert_w - A Fortran routine for calculating real 
+        !    values of the W-function,
+        !    ACM Transactions on Mathematical Software,
+        !    Volume 21, Number 2, June 1995, pages 172-181.
+        !
+        !  Input:
+        !
+        !    real ( kind = rk ) x: the argument.
+        !
+        !    integer nb: indicates the desired branch.
+        !    * 0, the upper branch;
+        !    * nonzero, the lower branch.
+        !
+        !    integer l: indicates the interpretation of X.
+        !    * 1, X is actually the offset from -(exp-1), so compute W(X-exp(-1)).
+        !    * not 1, X is the argument; compute W(X);
+        !
+        !  Output:
+        !
+        !    real ( kind = rk ) lambert_w: the approximate value of W(X).
+        !
+        implicit none
+        
+        integer, parameter :: rk = kind ( 1.0D+00 )
+        
+        real ( kind = rk ) an2
+        real ( kind = rk ) an3
+        real ( kind = rk ) an4
+        real ( kind = rk ) an5
+        real ( kind = rk ) an6
+        real ( kind = rk ) c13
+        real ( kind = rk ) c23
+        real ( kind = rk ) d12
+        real ( kind = rk ) delx
+        real ( kind = rk ) em
+        real ( kind = rk ) em2
+        real ( kind = rk ) em9
+        real ( kind = rk ) eta
+        integer i
+        integer l
+        real ( kind = rk ) lambert_w
+        integer nb
+        integer nbits
+        integer niter
+        real ( kind = rk ) reta
+        real ( kind = rk ) s2
+        real ( kind = rk ) s21
+        real ( kind = rk ) s22
+        real ( kind = rk ) s23
+        real ( kind = rk ) t
+        real ( kind = rk ) tb
+        real ( kind = rk ) tb2
+        real ( kind = rk ) temp
+        real ( kind = rk ) temp2
+        real ( kind = rk ) ts
+        real ( kind = rk ) x
+        real ( kind = rk ) x0
+        real ( kind = rk ) x1
+        real ( kind = rk ) xx
+        real ( kind = rk ) zl
+        real ( kind = rk ) zn
+        
+        lambert_w = - huge ( 1.0D+00 )
+        !
+        !  Various mathematical constants.
+        !
[…]
