+++ unit_tests/kerneltest.F90
@@ -1,26 +1,34 @@
 program testkernel
 
+!! -------------------------------------------------------------------------------------------
+!! The program produces the output of coagulation with the constant, linear and product kernel.
+!! Also has a routine to test the fragmentation as detailed in Lombart et al 2024
+!! -------------------------------------------------------------------------------------------
+
     use types
     use hdf5output
     use hdf5
     use parameters, only: path, datadir, read_parameters
+    use util_functions, only: lambert_w
 
     implicit none
     !  ------- set parameters here --------------------------------------------------------------------------------------
-   integer, parameter             :: ntot = 1000      ! total number of representative particles (swarms)
+   integer, parameter             :: ntot = 5e4     ! total number of representative particles (swarms)
    real, parameter                :: dens = 1.       ! dust volume mass density
    real, parameter                :: mswrm = 10.e20  ! mass of the swarm
 #ifdef FRAGTEST
-   real, parameter                :: m0 = 1e-6
-   real, parameter                :: gamma = 1e4
+   real, parameter                :: m0 = 1e-6      ! minimum mass of the particles
+   real, parameter                :: gamma = 1e3     ! distrubution parameter
+   real, parameter                :: mmax = 1e3      ! maximum mass of the particles  
+   integer, parameter             :: repeat = 1      ! repeat is set to 1 for the fragmentation simulation
 #else
    real, parameter                :: m0 = 1.         ! mass of the monomer (initial mass of the particles)
+   integer, parameter             :: repeat = 10
 #endif
    real, parameter                :: tend = 20.      ! maximum time of the simulation
    real, parameter                :: dtout = 4.      ! time step for the output
    integer, parameter             :: nbins = 250     ! number of bins for output histograms
-   integer, parameter             :: repeat = 10     ! number of times you want to repeat a simulation
-   real, parameter                :: dmmax = 0.001
+   real, parameter                :: dmmax = 0.001   ! collision optimisation parameter
    ! -------------------------------------------------------------------------------------------------------------------
 
    type(swarm), dimension(ntot)      :: swarms             ! list of the swarms
@@ -47,35 +55,43 @@ program testkernel
    real, dimension(:), allocatable :: t_arr
    real, dimension(nbins)         :: m2fm              ! mass distribution function
    real, dimension(nbins+1)       :: mgrid             ! mass grid
-   integer                        :: k = 1
+   integer                        :: k = 1, iter = 0
    real                           :: ll, lll, nord
+   real(kind=16)                           :: C,arg
 
-  datadir = 'kernels'
+  !datadir = 'kernels'
   call get_command_argument(1, ctrl_file)
   write(*,*) 'Reading parameters...'
   call read_parameters(ctrl_file)
 
 #ifdef LINEARKERNEL
-   !dtime = 20.
    allocate(t_arr(5))
    t_arr = (/4., 8., 12., 16., 20./)
 #elif PRODUCTKERNEL
-   !dtime = 0.9
    allocate(t_arr(3))
    t_arr = (/0.4, 0.7, 0.9/)
+#elif FRAGTEST
+   allocate(t_arr(10))
+   t_arr(1)=0.
+   do i=2,10
+      t_arr(i) = t_arr(i-1) + 3.33e-4
+   enddo
 #else
-   !dtime = 100000.
    allocate(t_arr(6))
[…]

[…]
