+++ paper/paper.bib
@@ -369,7 +369,7 @@ journal = {The Astrophysical Journal},
 	pages = "215-232",
 }
 
-@article{ Flores-Rivera2025,
+@article{ Floresrivera2025,
 	author = {{Flores-Rivera, Lizxandra} and {Lambrechts, Michiel} and {Gavino, Sacha} and {Lorek, Sebastian} and {Flock, Mario} and {Johansen, Anders} and {Mignone, Andrea}},
 	title = {UV processing of icy pebbles in the outer parts of VSI-turbulent disks},
 	DOI= "10.1051/0004-6361/202452933",
