+++ paper/paper.bib
@@ -294,7 +294,7 @@ eprint = {
 
 }
 
-@article{Okuzumi_2009, 
+@article{Okuzumi2009, 
 doi = {10.1088/0004-637X/707/2/1247}, 
 url = {https://doi.org/10.1088/0004-637X/707/2/1247}, 
 year = {2009}, 
@@ -308,7 +308,7 @@ title = {NUMERICAL MODELING OF THE COAGULATION AND POROSITY EVOLUTION OF DUST AG
 journal = {The Astrophysical Journal},  
 } 
 
-@article{Gurrutxaga_2026,
+@article{Gurrutxaga2026,
 doi = {10.3847/1538-4357/ae6104},
 url = {https://doi.org/10.3847/1538-4357/ae6104},
 year = {2026},
@@ -323,7 +323,7 @@ journal = {The Astrophysical Journal},
 }
 
 @article{ Houge2025,
-	author = {{Houge, Adrien} and {Johansen, Anders} and {Bergin, Edwin} and {Ciesla, Fred J.} and {Bitsch, Bertram} and {Lambrechts, Michiel} and {Henning, Thomas} and {Perotti, Giulia}},
+	author = {{Houge}, Adrien and {Johansen, Anders} and {Bergin, Edwin} and {Ciesla, Fred J.} and {Bitsch, Bertram} and {Lambrechts, Michiel} and {Henning, Thomas} and {Perotti, Giulia}},
 	title = {Burned to ashes: How the thermal decomposition of refractory organics in the inner protoplanetary disc impacts the gas-phase C/O ratio},
 	DOI= "10.1051/0004-6361/202555164",
 	url= "https://doi.org/10.1051/0004-6361/202555164",
@@ -334,7 +334,7 @@ journal = {The Astrophysical Journal},
 }
 
 @article{ Pinilla2012,
-	author = {{Pinilla, P.} and {Birnstiel, T.} and {Ricci, L.} and {Dullemond, C. P.} and {Uribe, A. L.} and {Testi, L.} and {Natta, A.}},
+	author = {{Pinilla} P. and {Birnstiel, T.} and {Ricci, L.} and {Dullemond, C. P.} and {Uribe, A. L.} and {Testi, L.} and {Natta, A.}},
 	title = {Trapping dust particles in the outer regions of protoplanetary disks},
 	DOI= "10.1051/0004-6361/201118204",
 	url= "https://doi.org/10.1051/0004-6361/201118204",
@@ -346,7 +346,7 @@ journal = {The Astrophysical Journal},
 }
 
 @article{ Woitke2016,
-	author = {{Woitke, P.} and {Min, M.} and {Pinte, C.} and {Thi, W.-F.} and {Kamp, I.} and {Rab, C.} and {Anthonioz, F.} and {Antonellini, S.} and {Baldovin-Saavedra, C.} and {Carmona, A.} and {Dominik, C.} and {Dionatos, O.} and {Greaves, J.} and {G\"udel, M.} and {Ilee, J. D.} and {Liebhart, A.} and {M\'enard, F.} and {Rigon, L.} and {Waters, L. B. F. M.} and {Aresu, G.} and {Meijerink, R.} and {Spaans, M.}},
[…]
