+++ paper/paper.bib
@@ -308,7 +308,7 @@ title = {NUMERICAL MODELING OF THE COAGULATION AND POROSITY EVOLUTION OF DUST AG
 journal = {The Astrophysical Journal},  
 } 
 
-@article{Gurrutxaga2026,
+@article{Gurrutxaga2026a,
 doi = {10.3847/1538-4357/ae6104},
 url = {https://doi.org/10.3847/1538-4357/ae6104},
 year = {2026},
@@ -378,4 +378,30 @@ journal = {The Astrophysical Journal},
 	year = 2025,
 	volume = 693,
 	pages = "A281",
+}
+
+@article{ Gurrutxaga2026b,
+	author = {{Gurrutxaga, Nerea} and {Vaikundaraman, Vignesh} and {Drazkowska, Joanna}},
+	title = {A Monte Carlo method for tracking dust properties during coagulation in protoplanetary disks},
+	DOI= "10.1051/0004-6361/202659523",
+	url= "https://doi.org/10.1051/0004-6361/202659523",
+	journal = {A\&A},
+	year = 2026,
+	volume = 709,
+	pages = "A164",
+}
+
+@article{Dra̧żkowska2019,
+doi = {10.3847/1538-4357/ab46b7},
+url = {https://doi.org/10.3847/1538-4357/ab46b7},
+year = {2019},
+month = {nov},
+publisher = {The American Astronomical Society},
+volume = {885},
+number = {1},
+pages = {91},
+author = {Dra̧żkowska, Joanna and Li, Shengtai and Birnstiel, Til and Stammler, Sebastian M. and Li, Hui},
+title = {Including Dust Coagulation in Hydrodynamic Models of Protoplanetary Disks: Dust Evolution in the Vicinity of a Jupiter-mass Planet},
+journal = {The Astrophysical Journal},
+abstract = {Dust growth is often neglected when building models of protoplanetary disks due to its complexity and computational expense. However, it does play a major role in shaping the evolution of protoplanetary dust and planet formation. In this paper, we present a numerical model coupling 2D hydrodynamic evolution of a protoplanetary disk, including a Jupiter-mass planet, and dust coagulation. This is obtained by including multiple dust fluids in a single grid-based hydrodynamic simulation and solving the Smoluchowski equation for dust coagulation on top of solving for the hydrodynamic evolution. We find that fragmentation of dust aggregates trapped in a pressure bump outside of the planetary gap leads to an enhancement in the density of small grains. We compare the results obtained from the full-coagulation treatment to the commonly used, fixed-dust-size approach and to previously applied, less computationally intensive methods for including dust coagulation. We find that the full-coagulation results cannot be reproduced using the fixed-size treatment, but some can be mimicked using a relatively simple method for estimating the characteristic dust size in every grid cell.}
 }
