+++ paper/paper.bib
@@ -391,7 +391,7 @@ journal = {The Astrophysical Journal},
 	pages = "A164",
 }
 
-@article{Dra̧żkowska2019,
+@article{Drazkowska2019,
 doi = {10.3847/1538-4357/ab46b7},
 url = {https://doi.org/10.3847/1538-4357/ab46b7},
 year = {2019},
