+++ paper/paper.bib
@@ -334,7 +334,7 @@ journal = {The Astrophysical Journal},
 }
 
 @article{ Pinilla2012,
-	author = {{Pinilla} P. and {Birnstiel, T.} and {Ricci, L.} and {Dullemond, C. P.} and {Uribe, A. L.} and {Testi, L.} and {Natta, A.}},
+	author = {{Pinilla}, P. and {Birnstiel, T.} and {Ricci, L.} and {Dullemond, C. P.} and {Uribe, A. L.} and {Testi, L.} and {Natta, A.}},
 	title = {Trapping dust particles in the outer regions of protoplanetary disks},
 	DOI= "10.1051/0004-6361/201118204",
 	url= "https://doi.org/10.1051/0004-6361/201118204",
