### 2026-06-09 107953d
paper update
Corrected punctuation and grammatical errors in the paper's text for improved readability.
