+++ paper/paper.bib
@@ -381,7 +381,7 @@ journal = {The Astrophysical Journal},
 }
 
 @article{ Gurrutxaga2026b,
-	author = {{Gurrutxaga, Nerea} and {Vaikundaraman, Vignesh} and {Drazkowska, Joanna}},
+	author = {{Gurrutxaga}, Nerea and {Vaikundaraman, Vignesh} and {Drazkowska, Joanna}},
 	title = {A Monte Carlo method for tracking dust properties during coagulation in protoplanetary disks},
 	DOI= "10.1051/0004-6361/202659523",
 	url= "https://doi.org/10.1051/0004-6361/202659523",
