+++ paper/paper.md
@@ -26,7 +26,7 @@ bibliography: paper.bib
 
 # Summary
 
-The Solar System and other planetary systems are formed in a disk of gas and dust around the young forming central star called a protoplanetary disk. The micron-sized particles collide and grow through several processes to form the planetary architectures as we see them today. The overall gas dynamics, the interaction between the gas and dust and the collisions between the dust particles leave an imprint on the formation pathways of the planets and, therefore, on life. Recent observations with telescopes such as ALMA [@Andrews_2018] show a diversity of dust structures in protoplanetary disks, and it is imperative that we have models that are able to describe the observations and, more generally, to help move forward towards the goal of an end-to-end planet formation theory. We present `mcdust`, a parallel Monte Carlo code aimed at modelling the evolution of dust in protoplanetary disks to better understand the dust size distributions and dust collisional evolution.
+The Solar System and other planetary systems are formed in a disk of gas and dust around the young forming central star, called a protoplanetary disk. The micron-sized particles collide and grow through several processes to form the planetary architectures as we see them today. The overall gas dynamics, the interaction between the gas and dust and the collisions between the dust particles leave an imprint on the formation pathways of the planets and, therefore, on life. Recent observations with telescopes such as ALMA [@Andrews_2018] show a diversity of dust structures in protoplanetary disks, and it is imperative that we have models that are able to describe the observations and, more generally, to help move forward towards the goal of an end-to-end planet formation theory. We present `mcdust`, a parallel Monte Carlo code aimed at modelling the evolution of dust in protoplanetary disks to better understand the dust size distributions and dust collisional evolution.
 
 # Statement of Need
 
@@ -34,7 +34,7 @@ Modelling dust coagulation and evolution is an essential part of understanding p
 
 Dust coagulation is generally modelled with the Smoluchowski equation [@Smoluchowski1916], an integro-differential equation. The drawback of the methods is that when we add a property to track (e.g, charge, composition, porosity), it adds a dimension to solving the Smoluchowski Equation [@Stammler2017;@Blum2006;@Okuzumi2009].
 
-Alternatively, one can model dust coagulation by performing Monte Carlo simulations. Monte Carlo methods are well-suited for stochastic processes and that makes it a suitable method for dust coagulation [@Gillespie1975]. For such a method, the representative particle approach [@Zsom2008;@Ormel2007] is very suitable and is Lagrangian in nature, meaning we track the particles and hence their histories. We can add properties to the particles that can be tracked without much computational complexity. This advantage is very useful when we want to combine dust evolution and chemistry, where we want to look at the thermal histories of particles and their compositions when chemical models are included[@Floresrivera2025]. 
+Alternatively, one can model dust coagulation by performing Monte Carlo simulations. Monte Carlo methods are well-suited for stochastic processes, making them suitable for dust coagulation [@Gillespie1975]. For such a method, the representative particle approach [@Zsom2008;@Ormel2007] is very suitable and is Lagrangian in nature, meaning we track the particles and hence their histories. We can add properties to the particles that can be tracked without much computational complexity. This advantage is very useful when we want to combine dust evolution and chemistry, where we want to look at the thermal histories of particles and their compositions when chemical models are included [@Floresrivera2025]. 
 
 Although there have been several publications that have made use of the Monte Carlo method to model dust coagulation in protoplanetary disks [@Krijt2018;@Houge2023], there is no available open source code that models dust coagulation using Monte Carlo methods and such a code would be very helpful as an alternative way for modelling dust coagulation.
 
[…]
