+++ docs/source/index.rst
@@ -22,6 +22,5 @@
     notebooks/mcdust_example
     notebooks/kernel_tests
     example_1
-    notebooks/verticalcolumn
     citation
     
