### 2026-06-13 ee77be0
paper updates
