+++ paper/paper.md
@@ -32,7 +32,7 @@ The Solar System and other planetary systems are formed in a disk of gas and dus
 
 Modelling dust coagulation and evolution is an essential part of understanding protoplanetary disks and planet formation. The evolution of micrometer-sized dust grains to millimeter-sized aggregates sets the tone for planetesimal formation. Dust evolution is inextricably linked with the formation of substructures in disks [@Pinilla2012]. Dust dynamics is also of prime importance when understanding protoplanetary disk chemistry [@Woitke2016;@Houge2025]. Therefore, it is very important to understand the processes involved in dust growth and dynamics and how they influence different aspects of disk dynamics and planet growth in detail.
 
-Dust coagulation is generally modelled with the Smoluchowski equation [@Smoluchowski1916], an integro-differential equation. The drawback of the methods is that when we add a property to track (e.g, charge, composition, porosity), it adds a dimension to solving the Smoluchowski Equation [@Stammler2017;@Blum2006;@Okuzumi2009].
+Dust coagulation is generally modelled with the Smoluchowski equation [@Smoluchowski1916], an integro-differential equation. The drawback of the methods is that when we add a property to track (e.g, charge, composition, porosity), it adds a dimension to solving the Smoluchowski equation and necessisates a parameterized approach when considering dust properties to reduce the complexity in these methods.[@Stammler2017;@Blum2006;@Okuzumi2009].
 
 Alternatively, one can model dust coagulation by performing Monte Carlo simulations. Monte Carlo methods are well-suited for stochastic processes, making them suitable for dust coagulation [@Gillespie1975]. For such a method, the representative particle approach [@Zsom2008;@Ormel2007] is very suitable and is Lagrangian in nature, meaning we track the particles and hence their histories. We can add properties to the particles that can be tracked without much computational complexity. This advantage is very useful when we want to combine dust evolution and chemistry, where we want to look at the thermal histories of particles and their compositions when chemical models are included [@Floresrivera2025]. 
 
@@ -40,9 +40,9 @@ Although there have been several publications that have made use of the Monte Ca
 
 # State of the field
 
-Several codes exist to model dust growth in protoplanetary disks, and we focus on discussing the open-source codes in the field. There are semi-analytical methods like `twopoppy` [@Birnstiel2012] and `TriPoD` [@Pfeil2024] that model the overall dust size distributions without modelling full dust coagulation. Some examples of open source codes that model full dust coagulation are `dustpy` [@Stammler2022], a 1D code to simulate gas and dust evolution, and `cuDisc` [@Robinson2024], a 2D (radial-vertical) code modelling dust coagulation and disk evolution. Both of these codes solving the Smoluchowski equation are Eulerian in nature, i.e., they follow volume rather than mass. `PHANTOM` [@Price] is an open source Smoothed Particle Hydrodynamics (SPH) code that models dust growth using a monodisperse growth or a 'single-size' approximation and does not solve the Smoluchowski equation for dust growth [@Vericel2021]. 
[…]
