+++ paper/paper.bib
@@ -91,7 +91,7 @@ archivePrefix = {arXiv},
 
 
 @ARTICLE{Price,
-       author = {{Price}, Daniel J. and {Wurster}, James and {Tricco}, Terrence S. and {Nixon}, Chris and {Toupin}, St{\'e}ven and {Pettitt}, Alex and {Chan}, Conrad and {Mentiplay}, Daniel and {Laibe}, Guillaume and {Glover}, Simon and {Dobbs}, Clare and {Nealon}, Rebecca and {Liptai}, David and {Worpel}, Hauke and {Bonnerot}, Cl{\'e}ment and {Dipierro}, Giovanni and {Ballabio}, Giulia and {Ragusa}, Enrico and {Federrath}, Christoph and {Iaconi}, Roberto and {Reichardt}, Thomas and {Forgan}, Duncan and {Hutchison}, Mark and {Constantino}, Thomas and {Ayliffe}, Ben and {Hirsh}, Kieran and {Lodato}, Giuseppe},
+       author = {{Price}, D. J. and {Wurster}, J. and {Tricco}, T. S. and {Nixon}, C. and {Toupin}, S. and {Pettitt}, A. and {Chan}, C. and {Mentiplay}, D. and {Laibe}, G. and {Glover}, Simon and {Dobbs}, Clare and {Nealon}, Rebecca and {Liptai}, David and {Worpel}, Hauke and {Bonnerot}, Cl{\'e}ment and {Dipierro}, Giovanni and {Ballabio}, Giulia and {Ragusa}, Enrico and {Federrath}, Christoph and {Iaconi}, Roberto and {Reichardt}, Thomas and {Forgan}, Duncan and {Hutchison}, Mark and {Constantino}, Thomas and {Ayliffe}, Ben and {Hirsh}, Kieran and {Lodato}, G.},
         title = "{Phantom: A Smoothed Particle Hydrodynamics and Magnetohydrodynamics Code for Astrophysics}",
       journal = {Publications of the Astronomical Society of Australia},
      keywords = {accretion, accretion disks, hydrodynamics, ISM: general, magnetohydrodynamics (MHD), methods: numerical, Astrophysics - Instrumentation and Methods for Astrophysics, Astrophysics - Earth and Planetary Astrophysics, Astrophysics - Astrophysics of Galaxies, Astrophysics - High Energy Astrophysical Phenomena, Astrophysics - Solar and Stellar Astrophysics},
@@ -323,7 +323,7 @@ journal = {The Astrophysical Journal},
 }
 
 @article{ Houge2025,
-	author = {{Houge}, Adrien and {Johansen, Anders} and {Bergin, Edwin} and {Ciesla, Fred J.} and {Bitsch, Bertram} and {Lambrechts, Michiel} and {Henning, Thomas} and {Perotti, Giulia}},
+	author = {{Houge}, A. and {Johansen}, A. and {Bergin}, E. and {Ciesla}, F.J. and {Bitsch}, B. and {Lambrechts}, M. and {Henning}, T. and {Perotti}, G.},
 	title = {Burned to ashes: How the thermal decomposition of refractory organics in the inner protoplanetary disc impacts the gas-phase C/O ratio},
 	DOI= "10.1051/0004-6361/202555164",
 	url= "https://doi.org/10.1051/0004-6361/202555164",
@@ -381,7 +381,7 @@ journal = {The Astrophysical Journal},
 }
 
 @article{ Gurrutxaga2026b,
-	author = {{Gurrutxaga}, Nerea and {Vaikundaraman, Vignesh} and {Drazkowska, Joanna}},
+	author = {{Gurrutxaga}, N. and {Vaikundaraman}, V. and {Drazkowska}, J.},
[…]
