### 2026-06-13 e937330
script updates
