### 2026-06-13 63b969f
bib updates
