+++ paper/paper.bib
@@ -145,7 +145,7 @@ archivePrefix = {arXiv},
 }
 
 @article{ Vaikundaraman2025,
-	author = {{Vaikundaraman}, V. and {Drążkowska, Joanna} and {Binkert, Fabian} and {Birnstiel, Til} and {Miotello, Anna}},
+	author = {{Vaikundaraman}, V. and {Drążkowska}, Joanna and {Binkert}, Fabian and {Birnstiel}, Til and {Miotello}, Anna},
 	title = {Refractory carbon depletion by photolysis through dust collisions and vertical mixing},
 	DOI= "10.1051/0004-6361/202553850",
 	url= "https://doi.org/10.1051/0004-6361/202553850",
@@ -266,7 +266,7 @@ abstract = {It is believed that satellites of giant planets form in circumplanet
 }
 
 @article{Blum2006,
-author = {Jürgen Blum},
+author = {Blum, Jürgen},
 title = {Dust agglomeration},
 journal = {Advances in Physics},
 volume = {55},
@@ -370,7 +370,7 @@ journal = {The Astrophysical Journal},
 }
 
 @article{ Floresrivera2025,
-	author = {{Flores-Rivera}, Lizxandra and {Lambrechts, Michiel} and {Gavino, Sacha} and {Lorek, Sebastian} and {Flock, Mario} and {Johansen, Anders} and {Mignone, Andrea}},
+	author = {{Flores-Rivera}, L. and {Lambrechts}, Michiel and {Gavino}, Sacha and {Lorek}, Sebastian and {Flock}, Mario and {Johansen}, Anders and {Mignone}, Andrea},
 	title = {UV processing of icy pebbles in the outer parts of VSI-turbulent disks},
 	DOI= "10.1051/0004-6361/202452933",
 	url= "https://doi.org/10.1051/0004-6361/202452933",
