### 2026-06-13 08d1dc2
bib updates
