+++ paper/paper.bib
@@ -317,7 +317,7 @@ publisher = {The American Astronomical Society},
 volume = {1003},
 number = {2},
 pages = {132},
-author = {Gurrutxaga, Nerea and Drążkowska, Joanna and Vaikundaraman, Vignesh and Kleine, Thorsten},
+author = {{Gurrutxaga}, Nerea and {Drążkowska}, Joanna and {Vaikundaraman}, Vignesh and {Kleine}, Thorsten},
 title = {Carbonaceous Chondrites Provide Evidence for Late-stage Planetesimal Formation in a Pressure Bump},
 journal = {The Astrophysical Journal},
 }
