### 2026-06-13 1814150
bib updates
