### 2026-06-13 79796bc
bib updates
