+++ paper/paper.bib
@@ -211,7 +211,7 @@ abstract = {Current models of (exo)planet formation often rely on a large influx
 }
 
 @article{Houge2023,
-    author = {Houge, Adrien and Krijt, Sebastiaan},
+    author = {Houge, A. and Krijt, S.},
     title = {Collisional evolution of dust and water ice in protoplanetary discs during and after an accretion outburst},
     journal = {Monthly Notices of the Royal Astronomical Society},
     volume = {521},
@@ -317,7 +317,7 @@ publisher = {The American Astronomical Society},
 volume = {1003},
 number = {2},
 pages = {132},
-author = {{Gurrutxaga}, N. and {Drążkowska}, J. and {Vaikundaraman}, V. and {Kleine}, T.},
+author = {{Gurrutxaga}, N. and {Drążkowska}, J. and {Vaikundaraman}, V. and {Kleine}, T.s},
 title = {Carbonaceous Chondrites Provide Evidence for Late-stage Planetesimal Formation in a Pressure Bump},
 journal = {The Astrophysical Journal},
 }
