### 2026-06-22 98ed76e
Update issue templates
