+++ .github/ISSUE_TEMPLATE/custom.md
@@ -1,7 +1,7 @@
 ---
 name: Custom issue template
 about: Describe this issue template's purpose here.
-title: ''
+title: "[CUSTOM]"
 labels: ''
 assignees: ''
 
