+++ docs/source/bugs.rst
@@ -1,0 +1,23 @@
+Contribute -  Bugs, Features
+============================
+
+Is it a bug? - bug reports
+++++++++++++++++++++++++++
+
+If you encounter something odd/wrong while using `mcdust`, we appreciate if you can open a `bug report issue<https://github.com/vicky1997/mcdust/issues/new?template=bug_report.md>`` describing the issue based on the template. 
+
+
+
+
+
+Is it a feature? - feature reports
+++++++++++++++++++++++++++++++++++
+
+If you think `mcdust` can be extended with a feature or there is something missing in `mcdust`, we encourage you to open a `feature report issue<https://github.com/vicky1997/mcdust/issues/new?template=bug_report.md>`.
+
+
+
+Other contributions
++++++++++++++++++++
+
+We appreciate it if you want to contribute to mcdust. To contribute, please open a `pull request<https://github.com/vicky1997/mcdust/pulls>`` with the details to the changes you want to introduce to the software.
