+++ docs/source/bugs.rst
@@ -4,7 +4,7 @@ Contribute -  Bugs, Features
 Is it a bug? - bug reports
 ++++++++++++++++++++++++++
 
-If you encounter something odd/wrong while using `mcdust`, we appreciate if you can open a `bug report issue<https://github.com/vicky1997/mcdust/issues/new?template=bug_report.md>`` describing the issue based on the template. 
+If you encounter something odd/wrong while using `mcdust`, we appreciate if you can open a `bug report issue<https://github.com/vicky1997/mcdust/issues/new?template=bug_report.md>`_ describing the issue based on the template. 
 
 
 
@@ -13,11 +13,11 @@ If you encounter something odd/wrong while using `mcdust`, we appreciate if you 
 Is it a feature? - feature reports
 ++++++++++++++++++++++++++++++++++
 
-If you think `mcdust` can be extended with a feature or there is something missing in `mcdust`, we encourage you to open a `feature report issue<https://github.com/vicky1997/mcdust/issues/new?template=bug_report.md>`.
+If you think `mcdust` can be extended with a feature or there is something missing in `mcdust`, we encourage you to open a `feature report issue<https://github.com/vicky1997/mcdust/issues/new?template=bug_report.md>`_.
 
 
 
 Other contributions
 +++++++++++++++++++
 
-We appreciate it if you want to contribute to mcdust. To contribute, please open a `pull request<https://github.com/vicky1997/mcdust/pulls>`` with the details to the changes you want to introduce to the software.
+We appreciate it if you want to contribute to mcdust. To contribute, please open a `pull request<https://github.com/vicky1997/mcdust/pulls>`_ with the details to the changes you want to introduce to the software.
