### 2026-06-30 c313e2e
Fix grammatical errors and improve clarity in paper
