+++ paper/paper.md
@@ -26,25 +26,25 @@ bibliography: paper.bib
 
 # Summary
 
-The Solar System and other planetary systems are formed in a disk of gas and dust around the young forming central star, called a protoplanetary disk. The micron-sized particles collide and grow through several processes to form the planetary architectures as we see them today. The overall gas dynamics, the interaction between the gas and dust and the collisions between the dust particles leave an imprint on the formation pathways of the planets and, therefore, on life. Recent observations with telescopes such as ALMA [@Andrews_2018] show a diversity of dust structures in protoplanetary disks, and it is imperative that we have models that are able to describe the observations and, more generally, to help move forward towards the goal of an end-to-end planet formation theory. We present `mcdust`, a parallel Monte Carlo code aimed at modelling the evolution of dust in protoplanetary disks to better understand the dust size distributions and dust collisional evolution.
+The Solar System and other planetary systems are formed in a disk of gas and dust around the young, forming central star, called a protoplanetary disk. The micron-sized particles collide and grow through several processes to form the planetary architectures as we see them today. The overall gas dynamics, the interaction between the gas and dust, and the collisions between the dust particles leave an imprint on the formation pathways of the planets and, therefore, on life. Recent observations with telescopes such as ALMA [@Andrews_2018] show a diversity of dust structures in protoplanetary disks, and it is imperative that we have models that are able to describe the observations and, more generally, to help move forward towards the goal of an end-to-end planet formation theory. We present `mcdust`, a parallel Monte Carlo code aimed at modelling the evolution of dust in protoplanetary disks to better understand the dust size distributions and dust collisional evolution.
 
 # Statement of Need
 
 Modelling dust coagulation and evolution is an essential part of understanding protoplanetary disks and planet formation. The evolution of micrometer-sized dust grains to millimeter-sized aggregates sets the tone for planetesimal formation. Dust evolution is inextricably linked with the formation of substructures in disks [@Pinilla2012]. Dust dynamics is also of prime importance when understanding protoplanetary disk chemistry [@Woitke2016;@Houge2025]. Therefore, it is very important to understand the processes involved in dust growth and dynamics and how they influence different aspects of disk dynamics and planet growth in detail.
 
-Dust coagulation is generally modelled with the Smoluchowski equation [@Smoluchowski1916], an integro-differential equation. The drawback of the methods is that when we add a property to track (e.g, charge, composition, porosity), it adds a dimension to solving the Smoluchowski equation and necessisates a parameterized approach when considering dust properties to reduce the complexity in these methods.[@Stammler2017;@Blum2006;@Okuzumi2009].
+Dust coagulation is generally modelled with the Smoluchowski equation [@Smoluchowski1916], an integro-differential equation. The drawback of the methods is that when we add a property to track (e.g, charge, composition, porosity), it adds a dimension to solving the Smoluchowski equation and necessitates a parameterized approach when considering dust properties to reduce the complexity in these methods.[@Stammler2017;@Blum2006;@Okuzumi2009].
 
[…]
