### 2026-06-30 514cf6d
Revised paper draft to adhere to word count
