### 2026-06-30 74d256e
Refine text in paper to adhere to word count
