+++ paper/paper.md
@@ -30,17 +30,17 @@ The Solar System and other planetary systems are formed in a disk of gas and dus
 
 # Statement of Need
 
-Modelling dust coagulation and evolution is an essential part of understanding protoplanetary disks and planet formation. The evolution of micrometer-sized dust grains to millimeter-sized aggregates sets the tone for planetesimal formation. Dust evolution is inextricably linked with the formation of substructures in disks [@Pinilla2012]. Dust dynamics is also of prime importance when understanding protoplanetary disk chemistry [@Woitke2016;@Houge2025]. Therefore, it is important to understand the processes involved in dust growth and dynamics and how they influence different aspects of disk dynamics and planet growth in detail.
+Modelling dust coagulation and evolution is an essential part of understanding protoplanetary disks and planet formation. The evolution of micrometer-sized dust grains to millimeter-sized aggregates sets the tone for planetesimal formation. Dust evolution is inextricably linked with the formation of substructures in disks [@Pinilla2012] and is of prime importance when understanding disk chemistry [@Woitke2016;@Houge2025]. Therefore, it is important to understand the processes involved in dust growth and dynamics and how they influence different aspects of disk dynamics and planet growth in detail.
 
 Dust coagulation is generally modelled with the Smoluchowski equation [@Smoluchowski1916], an integro-differential equation. The drawback of the methods is that when we add a property to track (e.g, charge, composition, porosity), it adds a dimension to solving the Smoluchowski equation and necessitates a parameterized approach when considering dust properties to reduce the complexity in these methods.[@Stammler2017;@Blum2006;@Okuzumi2009].
 
 One can also model dust coagulation by performing Monte Carlo simulations. Monte Carlo methods are well-suited for stochastic processes, making them suitable for dust coagulation [@Gillespie1975]. For such a method, the representative particle approach [@Zsom2008;@Ormel2007] is very suitable, allowing us to track the particles and hence their histories. We can add properties to the particles that can be tracked without much computational complexity. This advantage is very useful when we want to combine dust evolution and chemistry, allowing us to examine the thermal histories of particles and their compositions when chemical models are included [@Floresrivera2025]. 
 
-Although there have been several publications that have made use of the Monte Carlo method to model dust coagulation in protoplanetary disks [@Krijt2018;@Houge2023], there is no available open-source code that models dust coagulation using Monte Carlo methods, and such a code would be very helpful to model dust coagulation.
+Although there have been several publications that have made use of the Monte Carlo method to model dust coagulation in protoplanetary disks [@Krijt2018;@Houge2023], there is no available open-source code that models dust coagulation using Monte Carlo methods, and such a code would be very helpful.
 
 # State of the field
 
[…]
