### 2026-07-03 #10140 @editorialbot
:wave: @openjournals/aass-eics, this paper is ready to be accepted and published.
